requirements.txt
appnope==0.1.0
asgiref==3.2.10
backcall==0.2.0
beautifulsoup4==4.8.2
boto3==1.7.84
botocore==1.10.84
buckup==0.1a2
certifi==2020.4.5.2
chardet==3.0.4
decorator==4.4.2
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.7
django-cogwheels==0.3
django-extensions==2.2.9
django-modelcluster==5.0.2
django-social-share==1.4.0
django-storages==1.9.1
django-taggit==1.3.0
django-toolbelt==0.0.1
django-treebeard==4.3.1
djangorestframework==3.11.0
docutils==0.16
draftjs-exporter==2.1.7
gunicorn==20.0.4
html5lib==1.0.1
icalendar==4.0.6
idna==2.9
ipython-genutils==0.2.0
isoweek==1.3.3
jedi==0.17.1
jmespath==0.10.0
l18n==2018.5
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.5
psycopg2==2.8.5
ptyprocess==0.6.0
Pygments==2.6.1
python-dateutil==2.8.0
pytz==2020.1
requests==2.24.0
s3transfer==0.1.13
sentry-sdk==0.15.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
static3==0.7.0
traitlets==4.3.3
Unidecode==1.1.1
urllib3==1.25.9
wagtail==2.9
wagtail-events==0.4.0
wagtailgmaps==1.0.1
wagtailmenus==3.0.2
wcwidth==0.2.4
webencodings==0.5.1
whitenoise==5.1.0
Willow==1.3
XlsxWriter==1.2.9
